[SW Update] Manage ECC errors to prevent bad synchronization
Check for unrecoverable ECC errors during the synchronization and
abort if active or dual returns some ECC errors.
If an ECC error is found while le_fwupdate_InitDownload(), the
mode UPDATE without SYNC is enabled to allow a whole update of
the dual.
Recompute the source checksum for UBI partition. Redo a try if
the checksum does not match to prevent against UBI access by
UBI driver while synchronizing.

Resolves: LE-7699
Change-Id: I72b12d3de8c69f3795999858f9437c84ea187cb3