[DOTA] Implement bsPatch for RAW delta update
RAW partitions (like kernel) are patched with bspatch when
updating by delta.
In case of DOTA images, the checksum and length of an image is
broken by the delta patch. This will require two new services
partition_OpenSwifotaPartition() and partition_CloseSwifotaPartition()
to prevent any close performed by the partition low layer according
to an unexpected size or rejected due to a bad checksum.
Add 'META' in CWE header type list.

Resolves: LE-11047
Change-Id: Ide55a5957828b9db346cef2636a1b574aff2381d