[fwupdate] Sometimes markgood is failed
Add a parameter onlyChkValidUbiData to function of
partition_CheckData().
If it is an UBI empty block, we calculate CRC32 with
the first page data only. Else if it is an UBI block,
we calculate CRC32 with the data from the first page
to the last page that is not full of 0xff. Else we
calculate CRC32 with all the block data.

Resolves: LE-11156
Related: LEAR7M-401
Change-Id: I9a3c629778b3dd905accdb1672444fa407eac35c
Signed-off-by: Anson Ouyang <aouyang@sierrawireless.com>