[PA_FLASH_UBI] PA flash UBI may destroy valid volume blocks
In case of bad blocks, pa_flash_Erase() will mark the block bad and
choose the next one. This is coherent when a flash partition contains
only images, but not when its supports a file system like ubi-fs or
squashfs or is an ubi container. In that case, the next block to use
must be chosen by the UBI layer.

Resolves: LE-12031
Change-Id: Ied9bdcb39423fe418970c596048977c457ecf7c9