[FWUPDATE] AVMS reports WDSI 15 after FOTA
When performing FOTA upgrade from the version before fwupdate PA
refactoring, to the version after refactoring, FOTA job returns
the failed result (although the firmware is actually upgraded).
The reason for failure is that new PA can't find the download
status file on the file system.
The fix is: if the download status file is not found, query the
modem the same way the system behaved before refactoring.

Resolves: LE-12382
Change-Id: Id19929a6cfa9b55bb679ee9c27bc5aeb8d061b0b